# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/link-sets": {
      "post": {
        "operationId": "createLinkSet",
        "summary": "Create a new Link Set",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkSetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The Link Set was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkSet"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateLinkSet",
        "summary": "Update an existing Link Set",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLinkSetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The Link Set was updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkSet"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getLinkSets",
        "summary": "Get all paginated Link Sets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Link Sets were found and returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkSet"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/link-sets/{id}": {
      "get": {
        "operationId": "getLinkSet",
        "summary": "Get a Link Set by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Link Set was found and returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkSet"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteLinkSet",
        "summary": "Delete a Link Set by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/external-resolvers": {
      "post": {
        "operationId": "createExternalResolver",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExternalResolverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "updateExternalResolver",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExternalResolverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "getManyExternalResolvers",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/external-resolvers/{id}": {
      "get": {
        "operationId": "getExternalResolver",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "deleteExternalResolver",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/*": {
      "get": {
        "operationId": "ResolverController_resolve",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/.well-known/resolver": {
      "get": {
        "operationId": "ResolverController_getResolverMetadata",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Hermes",
    "description": "Hermes Link Resolver",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UpsertLinkDto": {
        "type": "object",
        "properties": {
          "relationType": {
            "type": "string",
            "description": "The relation type of the link."
          },
          "href": {
            "type": "string",
            "description": "The href of the link."
          },
          "objectKey": {
            "type": "string",
            "description": "The object key of the link, if TYPE is OBJECT."
          },
          "title": {
            "type": "string",
            "description": "The title of the link."
          },
          "lang": {
            "description": "The language of the link.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "relationType",
          "href",
          "title"
        ]
      },
      "CreateLinkSetDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "The identifier of the Link Set."
          },
          "qualifier": {
            "type": "string",
            "description": "The qualifier of the Link Set."
          },
          "links": {
            "description": "The links of the Link Set.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpsertLinkDto"
            }
          }
        },
        "required": [
          "identifier",
          "qualifier",
          "links"
        ]
      },
      "Link": {
        "type": "object",
        "properties": {
          "relationType": {
            "type": "string",
            "description": "The relation type of the link."
          },
          "href": {
            "type": "string",
            "description": "The href of the link."
          },
          "objectKey": {
            "type": "string",
            "description": "The object key of the link, if TYPE is OBJECT."
          },
          "title": {
            "type": "string",
            "description": "The title of the link"
          },
          "lang": {
            "description": "The language of the link.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The creation date of the Link."
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The last update date of the Link."
          }
        },
        "required": [
          "relationType",
          "title",
          "createdAt",
          "updatedAt"
        ]
      },
      "LinkSet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the Link Set."
          },
          "identifier": {
            "type": "string",
            "description": "The identifier of the Link Set."
          },
          "qualifier": {
            "type": "string",
            "description": "The qualifier of the Link Set."
          },
          "links": {
            "description": "The links of the Link Set.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The creation date of the Link Set."
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The last update date of the Link Set."
          }
        },
        "required": [
          "qualifier",
          "links",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateLinkSetDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "The identifier of the Link Set."
          },
          "qualifier": {
            "type": "string",
            "description": "The qualifier of the Link Set."
          },
          "links": {
            "description": "The links of the Link Set.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpsertLinkDto"
            }
          },
          "id": {
            "type": "string",
            "description": "The ID of the Link Set. This causes an UPDATE, not a CREATE."
          }
        },
        "required": [
          "identifier",
          "qualifier",
          "links"
        ]
      },
      "CreateExternalResolverDto": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "description": "The href of the External Resolver."
          },
          "pattern": {
            "type": "string",
            "description": "The pattern of the External Resolver."
          },
          "qualifier": {
            "type": "string",
            "description": "The qualifier of the External Resolver."
          },
          "childExternalResolvers": {
            "description": "The creation data of the External Resolver.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "href",
          "pattern",
          "qualifier",
          "childExternalResolvers"
        ]
      },
      "UpdateExternalResolverDto": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string",
            "description": "The href of the External Resolver."
          },
          "pattern": {
            "type": "string",
            "description": "The pattern of the External Resolver."
          },
          "qualifier": {
            "type": "string",
            "description": "The qualifier of the External Resolver."
          },
          "childExternalResolvers": {
            "description": "The creation data of the External Resolver.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string",
            "description": "The ID of the External Resolver."
          }
        },
        "required": [
          "href",
          "pattern",
          "qualifier",
          "childExternalResolvers"
        ]
      }
    }
  }
}